requirements.txt
altair==5.4.1
anyio==4.4.0
appdirs==1.4.4
asgiref==3.8.1
async-timeout==5.0.1
asyncio==3.4.3
asyncpg==0.30.0
attr==0.3.1
attrs==23.2.0
Automat==0.8.0
azure-core==1.30.2
azure-storage-blob==12.20.0
bcrypt==4.2.1
bleach==5.0.1
blinker==1.4
boto==2.49.0
boto3==1.34.125
botocore==1.34.125
boxing==0.1.4
cachetools==5.3.3
certifi==2019.11.28
chardet==3.0.4
charset-normalizer==3.3.2
Click==7.0
cloud-init==24.3.1
colorama==0.4.6
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
defusedxml==0.7.1
distro==1.9.0
distro-info==0.23+ubuntu1.1
Django==3.2.25
django-annoying==0.10.6
django-cors-headers==3.6.0
django-csp==3.7
django-debug-toolbar==3.2.1
django-environ==0.10.0
django-extensions==3.1.0
django-filter==2.4.0
django-model-utils==4.1.1
django-ranged-fileresponse==0.1.2
django-rq==2.5.1
django-storages==1.12.3
django-user-agents==0.4.0
djangorestframework==3.13.1
drf-dynamic-fields==0.3.0
drf-flex-fields==0.9.5
drf-generators==0.3.0
entrypoints==0.3
exceptiongroup==1.2.1
expiringdict==1.2.2
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.0
google-auth==2.30.0
google-cloud-appengine-logging==1.4.3
google-cloud-audit-log==0.2.5
google-cloud-core==2.4.1
google-cloud-logging==3.10.0
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.1
gprof2dot==2022.7.29
grpc-google-iam-v1==0.13.0
grpcio==1.64.1
grpcio-status==1.64.1
h11==0.14.0
htmlmin==0.1.12
httpcore==1.0.5
httplib2==0.14.0
httpx==0.27.0
humansignal-drf-yasg==1.21.9
hyperlink==19.0.0
idna==2.8
ijson==3.2.3
importlib-metadata==1.5.0
incremental==16.10.1
inflection==0.5.1
isodate==0.6.1
Jinja2==2.10.1
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
label-studio==1.12.1
label-studio-converter==0.0.58
label-studio-tools==0.0.4
language-selector==0.1
launchdarkly-server-sdk==8.2.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
lxml==5.2.2
macaroonbakery==1.3.1
markdown-it-py==3.0.0
MarkupSafe==1.1.0
mdurl==0.1.2
more-itertools==4.2.0
narwhals==1.15.1
netifaces==0.10.4
nltk==3.6.7
numpy==1.24.4
oauthlib==3.1.0
openai==1.34.0
ordered-set==4.0.2
packaging==24.1
pandas==2.0.3
pillow==10.3.0
proto-plus==1.23.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pyarrow==17.0.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycups==1.9.73
pydantic==1.10.16
pydeck==0.9.1
pygments==2.18.0
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyRFC3339==1.1
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.9.0.post0
python-debian==0.1.36+ubuntu1.1
python-json-logger==2.0.4
pytz==2022.7.1
PyYAML==6.0.1
redis==3.5.3
regex==2024.5.15
requests==2.31.0
requests-unixsocket==0.2.0
rich==13.9.4
rq==1.10.1
rsa==4.9
rules==2.2
s3transfer==0.10.1
SecretStorage==2.3.1
semver==3.0.2
sentry-sdk==2.5.1
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
smmap==5.0.1
sniffio==1.3.1
sqlparse==0.5.0
ssh-import-id==5.10
streamlit==1.40.1
systemd-python==234
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
tqdm==4.66.4
Twisted==18.9.0
typing-extensions==4.12.2
tzdata==2024.1
ua-parser==0.18.0
ubuntu-advantage-tools==20.3
ufw==0.36
ujson==5.10.0
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.18
user-agents==2.2.0
wadllib==1.3.3
watchdog==4.0.2
webencodings==0.5.1
xmljson==0.2.0
zipp==1.0.0
zope.interface==4.7.1
